requirements.txt
flask
numpy
pypdf[full]
pdf2image
gunicorn
google-generativeai
python-dotenv
markdown